Many changes! Optimizations and bug fixes.
All classes and functions are now in python package and modules ("cr" package).
