+++ README.md
@@ -1,0 +1,18 @@
+# Central Report
+
+**Central Report is a cool project, designed to monitor very easily servers or computers.**
+
+## The idea
+
+We have many ideas to simplify server monitoring:
+
+* Very simple installation
+* Cross platform (Unix, Linux, Windows)
+* Cloud conception
+* Alerts (push, mail, etc…)
+* High customizable dashboard
+* For public or Pros
+
+
+## Current status
+**Central Report is just at the beginning state. We have a lot of work to do, but we're motivated!**
