--- centralreport/collectors/Collector.py
@@ -1,20 +1,0 @@
-#
-# CentralReport - Indev version
-#
-
-class Collector:
-
-    def get_infos(self):
-        raise NameError,'Method not implemented yet'
-
-    def get_cpu(self):
-        raise NameError,'Method not implemented yet'
-
-    def get_memory(self):
-        raise NameError,'Method not implemented yet'
-
-    def get_loadaverage(self):
-        raise NameError,'Method not implemented yet'
-
-    def get_disks(self):
-        raise NameError,'Method not implemented yet'
